refactor(stats): sql functions in link statistics repo
